Move package amidst.mojangapi.file.world.filter to amidst.mojangapi.world.filter
