+++ src/main/java/amidst/mojangapi/world/filter/Criterion.java
@@ -1,9 +1,11 @@
-package amidst.mojangapi.file.world.filter;
+package amidst.mojangapi.world.filter;
 
 import java.util.Set;
 
+import amidst.documentation.Immutable;
 import amidst.mojangapi.world.coordinates.Region;
 
+@Immutable
 public interface Criterion {
 	public String getName();
