+++ src/main/java/amidst/mojangapi/world/filter/CriterionOr.java
@@ -1,12 +1,14 @@
-package amidst.mojangapi.file.world.filter;
+package amidst.mojangapi.world.filter;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import amidst.documentation.Immutable;
 import amidst.mojangapi.world.coordinates.Region;
 
+@Immutable
 public class CriterionOr implements Criterion {
 	private String name;
 	private List<Criterion> criteria;
@@ -16,6 +18,7 @@ public class CriterionOr implements Criterion {
 		this.name = name;
 		criteria = new ArrayList<>(list);
 		minimum = min;
+		
 	}
 	
 	public CriterionOr(String name, List<Criterion> list) {
